pina

+++ ts-web/vue/src/views/_dev/libBase/opencv-start-page.vue
@@ -1,11 +1,11 @@
 <template>
   <div class="this-page">
     <div class="inputoutput">
-      <img id="imageSrc" alt="No Image" />
+      <img id="imageSrc" alt="No Image" class="w-80 h-80"/>
       <div class="caption">imageSrc <input type="file" id="fileInput" name="file" /></div>
     </div>
     <div class="inputoutput">
-      <canvas id="canvasOutput" ></canvas>
+      <canvas id="canvasOutput" class="w-80 h-80" ></canvas>
       <div class="caption">canvasOutput</div>
     </div>
   </div>
@@ -14,30 +14,52 @@
 <script setup lang="ts">
 import { onMounted, onBeforeUnmount } from 'vue'
 import { loadJsDll } from '@/components/common/miniEx/utils/jsDll'
+import { add } from 'ol/coordinate'
 // vue
 onMounted(() => {
   init()
+  addImg()
 })
 onBeforeUnmount(() => {})
 const $ = (id: string) => document.getElementById(id)
 const init = async () => {
+
+
+
   // 同步加载cv
   const OPENCV_URL = 'https://docs.opencv.org/3.4.5/opencv.js'
+  // const OPENCV_URL = './opencv-start-page/opencv.js'
   await loadJsDll(OPENCV_URL)
-  let imgElement = $('imageSrc')
-  let mat = cv.imread(imgElement)
+  $('imageSrc')!.innerHTML = 'ready.';
+ 
+}
+let  addImg =()=>{
+  let imgElement = $('imageSrc') as any;
+  let inputElement = $('fileInput');
+  if(!inputElement)return
 
-  let dataMat = mat.data
-  for (let index = 0; index < dataMat.length; index++) {
-    const element = dataMat[index]
-    // console.log(element)
-    if (index % 128 === 0 && index > 0) {
-      dataMat[index] = 255
+  inputElement.addEventListener('change', (e) => {
+    if(!imgElement)return
+    if(!e.target)return
+    imgElement.src = URL.createObjectURL(e.target.files[0]);
+  }, false);
+  imgElement.onload = function () {
+    let mat = cv.imread(imgElement);
+    
+    let dataMat = mat.data
+    for (let index = 0; index < dataMat.length; index++) {
+      const element = dataMat[index];
+      // console.log(element)
+      if(index%128===0&&index>0){
+        dataMat[index]=255
+      }
+      
     }
-  }
-  cv.imshow('canvasOutput', mat)
-
-  mat.delete()
+    cv.imshow('canvasOutput', mat);
+    
+    
+    mat.delete();
+  };
 }
 </script>
